Move product type label to its own folder
