+++ client/profile-wizard/steps/product-types/index.js
@@ -1,65 +1,22 @@
 /**
  * External dependencies
  */
-import { __, sprintf } from '@wordpress/i18n';
-import { Component, Fragment } from '@wordpress/element';
+import { __ } from '@wordpress/i18n';
+import { Component } from '@wordpress/element';
 import { compose } from '@wordpress/compose';
-import {
-	Button,
-	CheckboxControl,
-	FormToggle,
-	Tooltip,
-} from '@wordpress/components';
+import { Button, CheckboxControl, FormToggle } from '@wordpress/components';
 import { includes, filter, get } from 'lodash';
-import interpolateComponents from 'interpolate-components';
 import { withDispatch, withSelect } from '@wordpress/data';
 import { getSetting } from '@woocommerce/wc-admin-settings';
-import { H, Card, Link, Pill } from '@woocommerce/components';
+import { H, Card } from '@woocommerce/components';
 import { ONBOARDING_STORE_NAME } from '@woocommerce/data';
 import { recordEvent } from '@woocommerce/tracks';
 
 /**
  * Internal dependencies
  */
-import './product-types.scss';
-
-function getLabel( description, yearlyPrice, isMonthlyPricing ) {
-	if ( ! yearlyPrice ) {
-		return description;
-	}
-
-	const priceDescription = isMonthlyPricing
-		? sprintf(
-				__( '$%f per month', 'woocommerce-admin' ),
-				( yearlyPrice / 12.0 ).toFixed( 2 )
-		  )
-		: sprintf( __( '$%f per year', 'woocommerce-admin' ), yearlyPrice );
-
-	/* eslint-disable @wordpress/i18n-no-collapsible-whitespace */
-	const toolTipText = __(
-		"This product type requires a paid extension.\nWe'll add this to a cart so that\nyou can purchase and install it later.",
-		'woocommerce-admin'
-	);
-	/* eslint-enable @wordpress/i18n-no-collapsible-whitespace */
-
-	return (
-		<Fragment>
-			<span className="woocommerce-product-wizard__product-types__label">
-				{ description }
-			</span>
-			<Tooltip text={ toolTipText } position="bottom center">
-				<span>
-					<Pill>
-						<span className="screen-reader-text">
-							{ toolTipText }
-						</span>
-						{ priceDescription }
-					</Pill>
-				</span>
-			</Tooltip>
-		</Fragment>
-	);
-}
+import ProductTypeLabel from './label';
+import './style.scss';
 
 class ProductTypes extends Component {
 	constructor( props ) {
@@ -143,12 +100,6 @@ class ProductTypes extends Component {
 		);
 	}
 
-	onLearnMore( slug ) {
-		recordEvent( 'storeprofiler_store_product_type_learn_more', {
-			product_type: slug,
-		} );
-	}
-
 	render() {
 		const { productTypes = {} } = getSetting( 'onboarding', {} );
 		const { error, isMonthlyPricing, selected } = this.state;
@@ -168,46 +119,27 @@ class ProductTypes extends Component {
 				<Card>
 					<div className="woocommerce-profile-wizard__checkbox-group">
 						{ Object.keys( productTypes ).map( ( slug ) => {
-							const label = getLabel(
-								productTypes[ slug ].label,
-								productTypes[ slug ].yearly_price,
-								isMonthlyPricing
-							);
-							const moreUrl = productTypes[ slug ].more_url;
-							const helpText =
-								productTypes[ slug ].description &&
-								interpolateComponents( {
-									mixedString:
-										productTypes[ slug ].description +
-										( productTypes[ slug ].more_url
-											? ' {{moreLink/}}'
-											: '' ),
-									components: {
-										moreLink: moreUrl ? (
-											<Link
-												href={ moreUrl }
-												target="_blank"
-												type="external"
-												onClick={ () =>
-													this.onLearnMore( slug )
-												}
-											>
-												{ __(
-													'Learn more',
-													'woocommerce-admin'
-												) }
-											</Link>
-										) : (
-											''
-										),
-									},
-								} );
-
 							return (
 								<CheckboxControl
 									key={ slug }
-									label={ label }
-									help={ helpText }
+									label={
+										<ProductTypeLabel
+											description={
+												productTypes[ slug ].label
+											}
+											annualPrice={
+												productTypes[ slug ]
+													.yearly_price
+											}
+											isMonthlyPricing={
+												isMonthlyPricing
+											}
+											moreUrl={
+												productTypes[ slug ].more_url
+											}
+											slug={ slug }
+										/>
+									}
 									onChange={ () => this.onChange( slug ) }
 									checked={ selected.includes( slug ) }
 									className="woocommerce-profile-wizard__checkbox"
