+++ client/profile-wizard/steps/product-types/style.scss
